[AIRFLOW-6867] Decouple DagBag and TaskInstance
